Fix license headers
