+++ tools/figma-inspector/src/main.tsx
@@ -2,8 +2,9 @@
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
 
 import React, { useEffect, useState } from "react";
-import { listenTS, copyToClipboard } from "./utils/utils";
+import { listenTS } from "./utils/bolt-utils";
 import "./main.css";
+import { copyToClipboard } from "./utils/utils.js";
 
 export const App = () => {
     const [title, setTitle] = useState("");
