+++ tools/figma-inspector/vite.config.code.ts
@@ -1,5 +1,5 @@
-// Copyright © SixtyFPS GmbH <info@slint.dev>
-// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
+// Copyright © Hyper Brew LLC
+// SPDX-License-Identifier: MIT
 
 import { defineConfig } from "vite";
 import { figmaCodePlugin } from "vite-figma-plugin";
